send 404 response when API encounters unknown path
