+++ viseron/components/webserver/api/const.py
@@ -1,0 +1,3 @@
+"""API constants."""
+
+API_BASE = "/api/v1"
